Refac: Create organized test directory structure
- Add utilities/ folder with admin.py, exchange_client.py, etc.
- Add scripts/, examples/, manual/ folders for better organization
- Create utilities/__init__.py for easy imports
