+++ src/test/manual/websocket_tester/config.js
