+++ src/test/manual/websocket_tester/tradingdemo.yml
